refactored to Rate- and CommentSeller files
